requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication & JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client for OAuth
httpx==0.25.2

# MongoDB (for auth and client sections)
motor==3.3.2
pymongo==4.6.0
dnspython==2.4.2

# PostgreSQL (for purchase and expense modules)
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.0

# Date/time utilities
python-dateutil==2.8.2

# Environment and configuration
pydantic-settings==2.1.0 